Use existing source verison if possible
Previously, gimme did not look for a pre-existing installation of the
requested version if `GIMME_TYPE` was set to "source". This commit
places the environment file for a source version at a different
location from the corresponding binary version so that we can detect if
we already have a valid source installation.

This fixes https://github.com/travis-ci/gimme/issues/43.