gimme: Add version, arch params to _try_binary
To support installing a binary bootstrap compiler to be used for
building a cross compiler add "arch" and "version" parameters
to the _binary and _try_binary functions and replace any use of
${GIMME_ARCH} and ${GIMME_GO_VERSION} in those routines with the
passed parameters.

Also, to enhance readability, convert _binary and _try_binary to use
local variables for their passed parameters.

Signed-off-by: lupan2005 <lupan2005@gmail.com>
[Cleaned up commit message.]
Signed-off-by: Geoff Levand <geoff@infradead.org>