Ask MacOS to not back-up ~/.gimme/versions
The `~/.gimme/versions` dir is a cache, each item of which is presumed
immutable once present, and grows by hundreds of MB for each version
installed.  So it makes sense for developers using gimme to want to
exclude it from their system backups.

Add exclusion, with guidance commentary, structured to support easy
extension for more OSes.