semi-colon terminate lines for easier eval
Without semi-colons, you must write:

    eval "$(gimme 1.7.1)"

because lines run together.  With semi-colons at the end of each
statement, the need for those quotes mostly disappears.  They're
technically more correct and should be used in scripting, but they're
annoying for people testing out Go variants at the cmdline.  For the
constrained output of `gimme`, this is safe.  With this change, the
user can type:

    eval $(gimme 1.7.1)

and it works.

Fix the `.env` creation, and fix rendering of existing `.env` files by
inserting a semi-colon at the end of each line if not already present.

Tested with old and new envs both.

Signed-Off-By: Phil Pennock <pdp@spodhuis.org>