Be more robust to unexpected command-lines
In #79 a user encountered problems because they invoked as
`gimme install 1.7.4` which broke with an unhelpful diagnostic.

* Skip `install` as a no-op instead of treating as a version; we can no
  longer install-from-git a remote branch named `install`; we accept
  this limitation
* Force the `$GIMME_VERSION_PREFIX` to be an absolute path, by prefacing
  the current working directory if needed.  This avoids `tmutil`
  complaining that it was given a relative path.
* When checking that we were given a version, also check that it looks
  valid.  Valid is `\d+\.\d+.*` (as bash extglob pattern) for version
  numbers; if auto/git type, then if that didn't pass us, see if we can
  check it out.  If we can, then it's a valid ref, so accept it.  Else
  complain, saying explicitly which value we saw.

This does move when we might first invoke `_checkout` but this should
be harmless: the _checkout does _fetch which does a `mkdir -p` so the
same paths get forced to exist whether or not `~/.gimme` already exists.
Worst case scenario, it's a bad ref and so for a first-time user
`~/.gimme` springs into existence with git checked out but we fail and
don't exclude from backups, so their `~/.gimme` will be included in
backups until they invoke correctly.  I think that this is okay.