Paginate bucket in 'gimme stable'
This is a similar change to
https://github.com/travis-ci/gimme/pull/108/commits but in bash, rather
than python.

It's possible we should use only one implementation, not two, but this
seems like the simpler fix, especially since I'm not sure if depending
on python is appropriate for this script.

This fixes #111

Note, the 'maxResults' value is ignored if it's over 1000, per google
storage docs, so dropping it should have no behavioral change.