Implement parsing of git revisions specifiers
Implemented ability to work from git via specifiers; wondering if we
should switch to "must supply a `git:` prefix" for this, and perhaps for
branches/tags too?

We were double-pulling from git, and checking remote twice for URLs
which don't exist, and various other unpleasantness.  If the first pass
invocation via _assert_version_given said "yup, that git branch exists"
then for `GIMME_TYPE=auto` we'd try to fetch a tarball named for that
branch from the downloads page.

For the new stuff, we now canonicalize to a sane name for the version.
We also now use a regexp to restrict what we'll _try_ to pull from the
download mirrors, which helps both as belt+braces and because we don't
canonicalize tags/branches: I wanted to keep that behavior in case folks
depended upon it somehow.

We were also trusting the branch/tag to be sane to use in a filename in
`envs/` despite those both allowing `/` in them, so we now replace
forward-slashes in the filename to be written.

I've made `_checkout` avoid use of sub-shells and cleaned it up a fair
bit, I think it's easier to glance and see what's being tried now, even
if puzzling out the mechanism is "differently awkward", rather than
actively better.  (I think it's now easier to read for non-shell-expert
programmers with experience in most other languages.)

Added checks to runtests to make sure that all the things we expect to
resolve will resolve.