bootstrap: always host os/arch, latest branch-version
ASSUMPTION: gimme's bootstrap is not being used to make a Go release
which can be copied to another system and used _there_ to build, but is
instead invoked locally to build the final version, where the final
version will be used for any cross-compilation.  Therefore we always
want to use the `HOST` variants for OS/ARCH for the bootstrap stage.

So, use the new resolution stuff to consider 1.10.x, 1.9.x or 1.8.x for
cross-compiling only, where we'll grab the latest on the branch for
future cross-compiles rather than continue with something which might
have known code-generation bugs.

And always go for the ${GIMME_HOSTOS:?}.${GIMME_HOSTARCH:?} variants.

Local testing confined to "can install a bootstrap".

Might help with issue #42.