Clarify GIMME_TYPE shown to user; warn consistency
We showed messages such as `(using type 'auto')` and this led to
misinterpretation by folks not familiar with all of gimme's controls,
thinking that we'd built a version named `auto` instead of what they'd
asked for.

I believe, but have no UX evidence to back up, that this minor rewording
should make it clearer to read.

Also a couple of warn switches, for consistency, to make this a slightly
less trivial commit.