Handle 1.x as a pattern correctly
The previous logic worked for `1.NN.x` but not for `1.x`.  I had avoided
use of regular expressions based on input, as an attack vector.

Validate the input to be sane, then use it to construct a regex after
all, and search for proper releases.

We need a way to stub out the updates and have tests for this
functionality.

Fixes #152