Unset GO111MODULE within Gimme itself
Presumably when `GO111MODULE` is set in environ when calling `gimme`,
it's part of generic setup to be used by the Go toolchain when building
the code being CI tested and the intent is not to try to tell Go, when
building from source, that it should be requiring the new Go 1.11 module
semantics.

Forcibly unset `GO111MODULE` in Gimme itself, but do not do anything to
add this to the emitted environment variables.

Without this, `gimme master` currently fails; with this change, it
succeeds.