If GIMME_DEBUG >= 2 then `make.bash -v`
We already have logic to handle `GIMME_DEBUG` looking like a number and
capturing more output from `make.bash` here; adjust condition to make it
clearer which value triggers is, and also tell `make.bash` to be more
verbose and get more details in the captured log.

Having `make.bash -v`, before the the #159 unset of `GO111MODULE`, this
facility reveals all the `use of internal package` complaints which show
that building Go itself is incompatible with having `GO111MODULE` set in
environ.