Improved windows host support (git bash & cygwin)
When using windows as a host OS, GOROOT must be a Windows path.
This condition applies whenever invoking go, either as part of
the installation or when written into the .env file.  Also,
the .exe files in the Windows Go distribution do not have
an executable flag set.  This change addresses both of these
issues, allowing at least binary installs to run correctly when
running under git bash or cygwin.